Prototype of the class
This file is create a static class to defined the node, and create left and right variables which type is node.  In addition, after the static class, I also create some data fields, then create a no argument constructor and a constructor, in the constructor, I let the twenty random numbers input an array.